# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/message": {
            "get": {
                "description": "根据jobId分页获取消息列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "获取消息列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "jobId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "起始位置",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dao.ListMessagesResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "创建消息",
                "parameters": [
                    {
                        "description": "创建消息请求",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dao.CreateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dao.CreateMessageResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/message/{message_id}": {
            "get": {
                "description": "根据message_id获取消息详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "获取消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "消息message_id",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/dao.MessageSpec"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "消息不存在",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据message_id删除消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "删除消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "消息message_id",
                        "name": "message_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "消息不存在",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "内部服务器错误",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dao.CreateMessageRequest": {
            "type": "object",
            "required": [
                "jobId",
                "timestamp"
            ],
            "properties": {
                "detectBoxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.DetectionBox"
                    }
                },
                "imagePath": {
                    "type": "string"
                },
                "jobId": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "integer"
                },
                "videoPath": {
                    "type": "string"
                },
                "workflowResp": {
                    "$ref": "#/definitions/dao.WorkflowResp"
                }
            }
        },
        "dao.CreateMessageResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "dao.DetectionBox": {
            "type": "object",
            "properties": {
                "classId": {
                    "type": "integer"
                },
                "confidence": {
                    "type": "number"
                },
                "label": {
                    "type": "string"
                },
                "x1": {
                    "type": "integer"
                },
                "x2": {
                    "type": "integer"
                },
                "y1": {
                    "type": "integer"
                },
                "y2": {
                    "type": "integer"
                }
            }
        },
...
        "dao.ListMessagesResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.MessageSpec"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dao.MessageSpec": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "detectBoxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dao.DetectionBox"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "imagePath": {
                    "type": "string"
                },
                "jobId": {
                    "type": "integer"
                },
                "timestamp": {
                    "type": "string"
                },
                "videoPath": {
                    "type": "string"
                },
                "workflowResp": {
                    "$ref": "#/definitions/dao.WorkflowResp"
                }
            }
        },
...
        "dao.WorkflowResp": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                }
            }
        },
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        }
    }
}